+++ packages/native-federation-tests/src/configurations/hostPlugin.ts
@@ -12,9 +12,11 @@ const retrieveRemoteStringUrl = (remote: string) => {
   return splittedRemote[splittedRemote.length - 1];
 };
 
+const FILE_PROTOCOL = 'file:';
+
 const buildZipUrl = (hostOptions: Required<HostOptions>, remote: string) => {
   const remoteStringUrl = retrieveRemoteStringUrl(remote);
-  const remoteUrl = new URL(remoteStringUrl);
+  const remoteUrl = new URL(remoteStringUrl, FILE_PROTOCOL);
 
   const pathnameWithoutEntry = remoteUrl.pathname
     .split('/')
@@ -22,7 +24,9 @@ const buildZipUrl = (hostOptions: Required<HostOptions>, remote: string) => {
     .join('/');
   remoteUrl.pathname = `${pathnameWithoutEntry}/${hostOptions.testsFolder}.zip`;
 
-  return remoteUrl.href;
+  return remoteUrl.protocol === FILE_PROTOCOL
+    ? remoteUrl.pathname
+    : remoteUrl.href;
 };
 
 const resolveRemotes = (hostOptions: Required<HostOptions>) => {